UPDATE RENDITION && FIX INT/DOUBLE INCOHERENCE:
The image is now destroyed then recreated prior rendering a new one,
instead of setting the background to black.

All the inconsistencies regarding double and int mathematical
operations are corrected.

+++ srcs/raycasting/dda_rays.c
@@ -14,21 +14,21 @@
 
 void	ft_hzt_init(t_cub_data *data, t_dda *ray)
 {
-	ray->inter_y = floor(data->player.y / 64) * 64;
+	ray->inter_y = floor(data->player.y / 64.0) * 64.0;
 	if (ray->facing_up == FALSE)
-		ray->inter_y += 64;
+		ray->inter_y += 64.0;
 	else
-		ray->inter_y -= 1;
+		ray->inter_y -= 1.0;
 	ray->inter_x = data->player.x + (ray->inter_y - data->player.y)
 		/ ray->tan_theta;
-	ray->step_y = 64;
+	ray->step_y = 64.0;
 	if (ray->facing_up == TRUE)
-		ray->step_y *= -1;
-	ray->step_x = 64 / ray->tan_theta;
-	if (ray->facing_left == TRUE && ray->step_x > 0)
-		ray->step_x *= -1;
-	else if (ray->facing_left == FALSE && ray->step_x < 0)
-		ray->step_x *= -1;
+		ray->step_y *= -1.0;
+	ray->step_x = 64.0 / ray->tan_theta;
+	if (ray->facing_left == TRUE && ray->step_x > 0.0)
+		ray->step_x *= -1.0;
+	else if (ray->facing_left == FALSE && ray->step_x < 0.0)
+		ray->step_x *= -1.0;
 	ray->hzt_x = ray->inter_x;
 	ray->hzt_y = ray->inter_y;
 }
@@ -36,7 +36,8 @@ void	ft_hzt_init(t_cub_data *data, t_dda *ray)
 void	ft_hzt_intersections(t_cub_data *data, t_dda *ray)
 {
 	ft_hzt_init(data, ray);
-	while (ray->hzt_x >= 0 && ray->hzt_x >> 6 < data->map_width && ray->hzt_y >= 0 && ray->hzt_y >> 6 < data->map_height)
+	while (ray->hzt_x >= 0 && ray->hzt_x >> 6 < data->map_width
+		&& ray->hzt_y >= 0 && ray->hzt_y >> 6 < data->map_height)
 	{
 		if (data->grid[ray->hzt_y >> 6][ray->hzt_x >> 6] == '1')
 		{
@@ -53,21 +54,21 @@ void	ft_hzt_intersections(t_cub_data *data, t_dda *ray)
 
 void	ft_vrt_init(t_cub_data *data, t_dda *ray)
 {
-	ray->inter_x = floor(data->player.x / 64) * 64;
+	ray->inter_x = floor(data->player.x / 64.0) * 64.0;
 	if (ray->facing_left == FALSE)
-		ray->inter_x += 64;
+		ray->inter_x += 64.0;
 	else
-		ray->inter_x -= 1;
+		ray->inter_x -= 1.0;
 	ray->inter_y = data->player.y + (ray->inter_x - data->player.x)
 		* ray->tan_theta;
-	ray->step_x = 64;
+	ray->step_x = 64.0;
 	if (ray->facing_left == TRUE)
-		ray->step_x *= -1;
-	ray->step_y = 64 * ray->tan_theta;
-	if (ray->facing_up == TRUE && ray->step_y > 0)
-		ray->step_y *= -1;
-	else if (ray->facing_up == FALSE && ray->step_y < 0)
-		ray->step_y *= -1;
+		ray->step_x *= -1.0;
+	ray->step_y = 64.0 * ray->tan_theta;
+	if (ray->facing_up == TRUE && ray->step_y > 0.0)
+		ray->step_y *= -1.0;
+	else if (ray->facing_up == FALSE && ray->step_y < 0.0)
+		ray->step_y *= -1.0;
 	ray->vrt_x = ray->inter_x;
 	ray->vrt_y = ray->inter_y;
 }
@@ -75,7 +76,8 @@ void	ft_vrt_init(t_cub_data *data, t_dda *ray)
 void	ft_vrt_intersections(t_cub_data *data, t_dda *ray)
 {
 	ft_vrt_init(data, ray);
-	while (ray->vrt_x >= 0 && ray->vrt_x >> 6 < data->map_width && ray->vrt_y >= 0 && ray->vrt_y >> 6 < data->map_height)
+	while (ray->vrt_x >= 0 && ray->vrt_x >> 6 < data->map_width
+		&& ray->vrt_y >= 0 && ray->vrt_y >> 6 < data->map_height)
 	{
 		if (data->grid[ray->vrt_y >> 6][ray->vrt_x >> 6] == '1')
 		{
@@ -96,10 +98,10 @@ void	ft_cast_ray(t_cub_data *data, t_dda *ray)
 	ft_set_dir(ray);
 	ft_hzt_intersections(data, ray);
 	ft_vrt_intersections(data, ray);
-	ray->hzt_dist = ft_dist_btw_pts((double)data->player.x,
-			(double)data->player.y, (double)ray->hzt_x, (double)ray->hzt_y);
-	ray->vrt_dist = ft_dist_btw_pts((double)data->player.x,
-			(double)data->player.y, (double)ray->vrt_x, (double)ray->vrt_y);
+	ray->hzt_dist = ft_dist_btw_pts(data->player.x,
+			data->player.y, ray->hzt_x, ray->hzt_y);
+	ray->vrt_dist = ft_dist_btw_pts(data->player.x,
+			data->player.y, ray->vrt_x, ray->vrt_y);
 	if (ray->vrt_dist < ray->hzt_dist)
 		ray->distance = ray->vrt_dist;
 	else if (ray->hzt_dist == ray->vrt_dist)
